+++ src/featureInfo/abstractFeatureInfoView.js
@@ -1,7 +1,6 @@
-import { VcsObject } from '@vcmap/core';
+import { renderTemplate, VcsObject } from '@vcmap/core';
 import { WindowSlot } from '../manager/window/windowManager.js';
 import { defaultTagOptions } from '../components/tables/VcsTable.vue';
-import { renderTemplate } from '../components/form-output/markdownHelper.js';
 
 /**
  * @typedef {Object} FeatureInfoProps